Changed How feed updates work
Now, the feed is loaded when the Activity is created (was in onResume before), and is automatically updated when the Activity is resumed if 5 minutes have passed. Before it would only update if you refreshed it manually, to conserve API calls, and avoid sending a request every time the Activity becomes visible. Auto reloading after at least 5 minutes is a good middle ground.